% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_areahex.R
\name{fo_area_hex}
\alias{fo_area_hex}
\title{Hexagon area}
\usage{
fo_area_hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r}{Lenght from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Numeric value
}
\description{
Function which calculates an area of a hexagon
based on its apexes.
}
\examples{
fo_area_hex(1,0,3, 45)

}
